// src/app/interfaces/ITest.ts
export interface ITest {
  idUser: string;
  urlReturn: string;
  idHeader: string;
  idTurn: string;
  documentType: string;
  documentNumber: string;
  account: string;
  source: string;
  min: string;
  SuscriptionState: string;
  ReasonState: string;
  Customer: string;
  OverDue: string;
  PlaceID: string;
  Type: string;
  originSource: string;
}